R6 class representing an evaluator.

An evaluator (or evaluation function) to evaluate the performance of methods in the Experiment.

Public fields

name

The name of the Evaluator.

eval_fun

The evaluation function.

eval_params

(Named) list of parameters to input into the evaluation function.

rmd_options

List of options to control the aesthetics of the displayed Evaluator's results table in the knitted R Markdown report. See prettyDT() for possible options.

show

If TRUE (default), show Evaluator's results as a table in the R Markdown report; if FALSE, hide output in the R Markdown report.

Methods

Public methods

Evaluator$new()

Evaluator$evaluate()

Evaluator$print()

Evaluator$clone()

Method new()

Create a new Evaluator.

Usage

Evaluator$new(eval_fun, name = NULL, rmd_options = list(), show = TRUE, ...)

Arguments

eval_fun

The evaluation function.

name

(Optional) The name of the Evaluator.

rmd_options

(Optional) List of options to control the aesthetics of the displayed Evaluator's results table in the knitted R Markdown report. See prettyDT() for possible options.

show

If TRUE (default), show Evaluator's results as a table in the R Markdown report; if FALSE, hide output in the R Markdown report.

...

Arguments to pass into eval_fun().

Details

When evaluating or running the Experiment (see Experiment$evaluate() or Experiment$run()), the named arguments fit_results and vary_params are automatically passed into the Evaluator function eval_fun() and serve as placeholders for the Experiment$fit() results (i.e., the results from the method fits) and the name of the varying parameter, respectively. To evaluate the performance of a method(s) fit then, the Evaluator function eval_fun() should almost always take in the named argument fit_results. See Experiment$fit() or fit_experiment() for details on the format of fit_results. If the Evaluator is used for Experiments with varying parameters, vary_params should be used as a stand in for the name of this varying parameter.

Returns

A new Evaluator object.

Method evaluate()

Evaluate the performance of method(s) in the Experiment using the Evaluator and its given parameters.

Usage

Evaluator$evaluate(fit_results, vary_params = NULL, ...)

Arguments

fit_results

A tibble, typically returned by the Experiment$fit() method.

vary_params

Name of parameters/arguments that were varied in the Experiment (see Experiment$get_vary_across()). Use NULL (default) if no vary_across component in Experiment run.

...

Not used.

Returns

Result of eval_fun(), coerced into a tibble.

Method print()

Print an Evaluator in a nice format, showing the Evaluator's name, function, parameters, and R Markdown options.

Usage

Evaluator$print()

Returns

The original Evaluator object.

Method clone()

The objects of this class are cloneable with this method.

Usage

Evaluator$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.