Create a simulation project

create_sim() initializes a directory for your simulation study. It wraps around usethis::create_project(), as well as usethis::use_git() and renv::init().

Usage

create_sim(path, init_git = TRUE, init_renv = FALSE, tests = TRUE, hpc = FALSE)

Arguments

path

A character specifying the path for your simulation directory.

init_git

A logical indicating whether to intialize your simulation directory as a git repository.

init_renv

A logical stating whether to initialize renv with init() from the renv package. Defaults to FALSE.

tests

A logical indicating whether to generate sub-directories for organizing unit tests. Defaults to TRUE.

hpc

A logical indicating whether to create sub-directories for organizing files related to high-power computing environments. Defaults to FALSE.

Examples

if (FALSE) { # \dontrun{
# create template directory for simulation project
create_sim("path/to/sim")} # }