Evaluate an Experiment.

Evaluate the performance of method(s) across all Evaluators in the Experiment and return results.

Usage

evaluate_experiment(
  experiment,
  fit_results,
  use_cached = FALSE,
  save = FALSE,
  verbose = 1,
  ...
)

Arguments

experiment

An Experiment object.

fit_results

A tibble, as returned by the fit method.

use_cached

Logical. If TRUE, find and return previously saved results. If cached results cannot be found, continue as if use_cached was FALSE.

save

If TRUE, save outputs to disk.

verbose

Level of verbosity. Default is 1, which prints out messages after major checkpoints in the experiment. If 2, prints additional debugging information for warnings and messages from user-defined functions (in addition to error debugging information). If 0, no messages are printed other than user-defined function error debugging information.

...

Not used.

Value

A list of evaluation result tibbles, one for each Evaluator.