Fix duplicate vary across parameter names.

Add "_dgp" or "_method" suffixes to parameter names that are found in both the DGP and Method vary across components. This is to avoid errors that occur from duplicate column names when trying to create a tibble.

fix_duplicate_param_names(
  dgp_params,
  method_params,
  duplicate_param_names = NULL
)

Arguments

dgp_params

A named list of the DGP parameters.

method_params

A named list of the Method parameters.

duplicate_param_names

A vector of parameter names that are varied across in both a DGP and a Method

Value

A named list of the DGP and Method parameters with no duplicate names.