Simulate linear response data.

Generate linear response data with a specified error distribution given the observed and unobserved design matrices.

generate_y_linear(
  X,
  U,
  betas = 0,
  betas_unobs = 0,
  intercept = 0,
  err = NULL,
  return_support = FALSE,
  ...
)

Arguments

X

Design data matrix of observed variables.

U

Design data matrix of unobserved (omitted) variables.

betas

Scalar or vector of coefficients corresponding to features in observed design matrix.

betas_unobs

Scalar or vector of coefficients corresponding to features in unobserved design matrix.

intercept

Scalar intercept term.

err

Function from which to generate simulate error vector. Default is NULL which adds no error to the DGP.

return_support

Logical specifying whether or not to return a vector of the support column names. If X has no column names, then the indices of the support are used.

...

Other arguments to pass to err() to generate the error vector.

Value

If return_support = TRUE, returns a list of two:

y

A response vector of length nrow(X).

support

A vector of feature indices indicating all features used in the true support of the DGP.

If return_support = FALSE, returns only the response vector y.

Examples

X <- generate_X_gaussian(n = 100, p = 2)
U <- generate_X_gaussian(n = 100, p = 2)

# generate the response from: y = 3*x_1 - x_2 + N(0, 1) errors
y <- generate_y_linear(X = X, betas = c(3, -1), err = rnorm)

# generate the response from: y = 3*x_1 - x_2 + u_1 + 2*u_2
y <- generate_y_linear(X = X, U = U, betas = c(3, -1), betas_unobs = c(1, 2))