Generate locally spiky smooth (LSS) response data.

Generate LSS response data with a specified error distribution given the observed data matrices.

generate_y_lss(
  X,
  k,
  s,
  thresholds = 1,
  signs = 1,
  betas = 1,
  intercept = 0,
  overlap = FALSE,
  err = NULL,
  return_support = FALSE,
  ...
)

Arguments

X
Data matrix or data frame.

k
Order of the interactions.

s
Number of interactions in the LSS model or a matrix of the support indices with each interaction taking a row in this matrix and ncol = k.

thresholds
A scalar or a s x k matrix of the thresholds for each term in the LSS model.

signs
A scalar or a s x k matrix of the sign of each interaction (1 means > while -1 means <).

betas
Scalar or parameter vector for interaction terms.

intercept
Scalar intercept term.

overlap
If TRUE, simulate support indices with replacement; if FALSE, simulate support indices without replacement (so no overlap)

err
Function from which to generate simulate error vector. Default is NULL which adds no error to the DGP.

return_support
Logical specifying whether or not to return a vector of the support column names. If X has no column names, then the indices of the support are used.

...
Other arguments to pass to err() to generate the error vector.

Value

If return_support = TRUE, returns a list of two:

y

A response vector of length nrow(X).

support

A vector of signed feature indices in the true (interaction) support of the DGP. For example, "1+_2-" means that the interaction between high values of feature 1 and low values of feature 2 appears in the underlying DGP.

If return_support = FALSE, returns only the response vector y.