Retrieve cached results from previously saved Experiment.

Read in cached results from disk from a previously saved Experiment.

Usage

get_cached_results(experiment, results_type, verbose = 0)

Arguments

experiment

An Experiment object.

results_type

Character string indicating the type of results to read in. Must be one of "experiment", "experiment_cached_params", "fit", "eval", or "viz".

verbose

Level of verbosity. Default is 1, which prints out messages after major checkpoints in the experiment. If 2, prints additional debugging information for warnings and messages from user-defined functions (in addition to error debugging information). If 0, no messages are printed other than user-defined function error debugging information.

Value

The cached results, specifically the cached Experiment object if results_type = "experiment", the cached fit results if results_type = "fit", the cached evaluation results if results_type = "eval", the cached visualization results if results_type = "viz", and the experiment parameters used in the cache if results_type = "experiment_cached_params".