Coerce list into a tibble.

Coerce a list into a tibble. Default is to coerce using tibble::as_tibble(), but if this fails, then coerce list into a tibble, where each non-scalar column in the tibble is of type list.

Usage

list_to_tibble(lst)

Arguments

lst

List to convert into a tibble.

Value

A tibble.