Create pretty datatable with custom bolding options.

Create pretty datatable with custom bolding options.

pretty_DT(
  X,
  digits = 3,
  sigfig = T,
  escape = F,
  rownames = TRUE,
  caption = "",
  na_disp = "NA",
  bold_function = NULL,
  bold_margin = NULL,
  bold_scheme = T,
  bold_color = NULL,
  options = list(),
  return_df = FALSE,
  ...
)

Arguments

X

Data frame or data matrix to display in table

digits

Number of digits to display for numeric values

sigfig

Logical. If TRUE, digits refers to the number of significant figures. If FALSE, digits refers to the number of decimal places.

escape

Logical. Whether or not to escape HTML entities in table. See DT::datatable() for details.

rownames

Logical. Whether or not to show rownames in table. See DT::datatable() for details.

caption

The table caption.

na_disp

Character string to display if NA entry is found in table.

bold_function

Optional function string or vector of function strings to use for bolding entries, e.g. ". == max(.)" or ". >= 0.5".

bold_margin

Specifies the margins of X that will be used to evaluate bold_function across, i.e., 0 = across entire matrix, 1 = across rows, and 2 = across columns. Required if bold_function = TRUE.

bold_scheme

Scalar or vector of logicals, indicating whether or not to apply bold_function to row if bold_margin = 1 and to column if bold_margin = 1, 2. Default is to apply bolding to all rows/columns.

bold_color

Color of bolded text.

options

See options argument in DT::datatable().

return_df

Logical. If TRUE, return data frame that was used as input into DT::datatable() in addition to the datatable output. If FALSE, only return the datatable output.

...

Additional arguments to pass to DT::datatable().

Value

If return_df = FALSE, returns a datatable object. Otherwise, returns a list of two:

dt

A datatable object.

df

A data frame that was used as input into DT::datatable().

Examples

# Show iris data table
pretty_DT(iris, caption = "Iris Data Table")

# Bold max value of each numeric column of Iris data in red
pretty_DT(iris, caption = "Iris Data Table",
         bold_function = ". == max(.)", bold_margin = 2,
         bold_scheme = c(TRUE, TRUE, TRUE, TRUE, FALSE), bold_color = "red")

            
# Bold min value of each row in Iris data
pretty_DT(iris %>% dplyr::select(-Species), 
         sigfig = TRUE, caption = "Iris Data Table",
         na_disp = "NA", bold_function = ". == min(.)", bold_margin = 1,
         bold_scheme = TRUE, bold_color = "black")