Arguments that are shared by multiple DGP library functions.

Arguments that are shared by multiple DGP library functions.

Arguments

data_split

Logical; if TRUE, splits data into training and test sets according to train_prop.

err

Function from which to generate simulate error vector. Default is NULL which adds no error to the DGP.

intercept

Scalar intercept term.

n

Number of samples.

p

Number of features.

return_support

Logical specifying whether or not to return a vector of the support column names. If X has no column names, then the indices of the support are used.

return_values

Character vector indicating what objects to return in list. Elements in vector must be one of "X", "y", "support".

support

Vector of feature indices in the true support of the DGP.

train_prop

Proportion of data in training set if data_split = TRUE.

X

Data matrix or data frame.

y

Response vector.

...

Other arguments to pass to err() to generate the error vector.

Value

A list of the named objects that were requested in return_values. See brief descriptions below.

X

A data.frame.

y

A response vector of length nrow(X).

support

A vector of feature indices indicating all features used in the true support of the DGP.

Note that if data_split = TRUE and "X", "y" are in return_values, then the returned list also contains slots for "Xtest" and "ytest".