Arguments that are shared by multiple Experiment helper funs.

Arguments that are shared by multiple Experiment helper funs.

Arguments

dgp
A DGP object.

evaluator
An Evaluator object.

experiment
An Experiment object.

future.globals
Character vector of names in the global environment to pass to parallel workers. Passed as the argument of the same name to future.apply::future_lapply and related functions. To set for all runs of the experiment, use the same argument during initialization.

future.packages
Character vector of packages required by parallel workers. Passed as the argument of the same name to future.apply::future_lapply and related functions. To set for all runs of the experiment, use the same argument during initialization.

future.seed
Passed as the argument of the same name in future.apply::future_apply.

method
A Method object.

n_reps
The number of replicates of the Experiment for this run.

parallel_strategy
A vector with some combination of "reps", "dgps", or "methods". Determines how computation will be distributed across available resources. Default is "reps".

save
If TRUE, save outputs to disk.

use_cached
Logical. If TRUE, find and return previously saved results. If cached results cannot be found, continue as if use_cached was FALSE.

verbose
Level of verbosity. Default is 1, which prints out messages after major checkpoints in the experiment. If 0, no messages are printed.

visualizer
A Visualizer object.

...
Not used.