Number of true positives

These functions calculate the tp() (number of true positives) of a measurement system compared to the reference results (the "truth").

Usage

tp(data, ...)

# S3 method for class 'data.frame'
tp(
  data,
  truth,
  estimate,
  estimator = NULL,
  na_rm = FALSE,
  case_weights = NULL,
  event_level = "first",
  ...
)

tp_vec(
  truth,
  estimate,
  estimator = NULL,
  na_rm = FALSE,
  case_weights = NULL,
  event_level = "first",
  ...
)

Arguments

data

Either a data.frame containing the columns specified by the truth and estimate arguments, or a table/matrix where the true class results should be in the columns of the table.

...

Not currently used.

truth

The column identifier for the true class results (that is a factor). This should be an unquoted column name although this argument is passed by expression and supports quasiquotation (you can unquote column names). For _vec() functions, a factor vector.

estimate

The column identifier for the predicted class results (that is also factor). As with truth this can be specified different ways but the primary method is to use an unquoted variable name. For _vec() functions, a factor vector.

estimator

One of: "binary", "macro", "macro_weighted", or "micro" to specify the type of averaging to be done. "binary" is only relevant for the two class case. The other three are general methods for calculating multiclass metrics. The default will automatically choose "binary" or "macro" based on estimate.

na_rm

A logical value indicating whether NA values should be stripped before the computation proceeds.

case_weights

The optional column identifier for case weights. This should be an unquoted column name that evaluates to a numeric column in data. For _vec() functions, a numeric vector, hardhat::importance_weights(), or hardhat::frequency_weights().

event_level

A single string. Either "first" or "second" to specify which level of truth to consider as the "event". This argument is only applicable when estimator = "binary". The default uses an internal helper that defaults to "first".

Value

A tibble with columns .metric, .estimator, and .estimate with 1 row of values.

For grouped data frames, the number of rows returned will be the same as the number of groups.

For tp_vec(), a single numeric value (or NA).

Examples

# Two class example data
two_class_example <- data.frame(
  truth = as.factor(sample(c("Class1", "Class2"), 100, replace = TRUE)),
  predicted = as.factor(sample(c("Class1", "Class2"), 100, replace = TRUE))
)

# Compute number of true positives
tp(two_class_example, truth = truth, estimate = predicted)
#> # A tibble: 1 × 3
#>   .metric .estimator .estimate
#>   <chr>   <chr>          <int>
#> 1 tp      binary            22
tp_vec(two_class_example$truth, two_class_example$predicted)
#> [1] 22