Varying across parameters in an Experiment.

Helper functions for adding, updating, removing, or getting a vary_across component in an Experiment. When a vary_across component is added and the Experiment is run, the Experiment is systematically varied across values of the specified parameter in the DGP or Method while all other parameters are held constant at their baseline value.

add_vary_across(.experiment, .dgp, .method, ...)

update_vary_across(.experiment, .dgp, .method, ...)

remove_vary_across(experiment, dgp, method, param_names = NULL)

get_vary_across(experiment)

Arguments

.experiment, experiment

An Experiment object.

.dgp, dgp

Name of DGP to vary in the Experiment. Can also be a DGP object that matches one in the Experiment.

.method, method

Name of Method to vary in the Experiment. Can also be a Method object that matches one in the Experiment.

...

Any number of named arguments where names match an argument in the user-specified DGP or Method function and values are vectors (for scalar parameters) or lists (for arbitrary parameters).

param_names

A character vector of parameter names to remove. If not provided, the entire set of vary_across parameters will be removed for the specified DGP/Method.

Value

In the case of get_vary_across, a nested list with entries "dgp" and "method" that contains the parameters to vary across for each DGP and Method in the Experiment. Otherwise, the original Experiment object passed to *_vary_across().

Details

One of the .dgp or .method arguments (but not both) must be provided when using add_vary_across() and update_vary_across. For remove_vary_across(), if both the dgp and method arguments are not provided, then all vary_across parameters from the experiment are removed.