Generate correlated Gaussian covariates and LSS response data.

Generate normally-distributed covariates that are potentially correlated and LSS response data with a specified error distribution.

correlated_lss_gaussian_dgp(
  n,
  p_uncorr,
  p_corr,
  s_uncorr = p_uncorr,
  s_corr = p_corr,
  corr,
  k,
  thresholds = 0,
  signs = 1,
  betas = 1,
  intercept = 0,
  overlap = FALSE,
  mixed_int = FALSE,
  err = NULL,
  data_split = FALSE,
  train_prop = 0.5,
  return_values = c("X", "y", "support"),
  ...
)

Arguments

n
Number of samples.

p_uncorr
Number of uncorrelated features.

p_corr
Number of features in correlated group.

s_uncorr
Number of interactions from features in uncorrelated group.

s_corr
Number of interactions from features in correlated group.

corr
Correlation between features in correlated group.

k
Order of the interactions.

thresholds
A scalar or a s x k matrix of the thresholds for each term in the LSS model.

signs
A scalar or a s x k matrix of the sign of each interaction (1 means > while -1 means <).

betas
Scalar or parameter vector for interaction terms.

intercept
Scalar intercept term.

overlap
If TRUE, simulate support indices with replacement; if FALSE, simualte support indices without replacement (so no overlap).

mixed_int
If TRUE, correlated and uncorrelated variables are mixed together when constructing an interaction of order-k. If FALSE, each interaction of order-k is composed of only correlated variables or only uncorrelated variables.

err
Function from which to generate simulate error vector. Default is NULL which adds no error to the DGP.

data_split
Logical; if TRUE, splits data into training and test sets according to train_prop.

train_prop
Proportion of data in training set if data_split = TRUE.

return_values
Character vector indicating what objects to return in list. Elements in vector must be one of "X", "y", "support".

...
Other arguments to pass to err() to generate the error vector.

Value

A list of the named objects that were requested in return_values. Note that if data_split = TRUE and "X", "y" are in return_values, then the returned list also contains slots for "Xtest" and "ytest".