Create documentation template for the R Markdown results report.

Create documentation template (a series of .md files) to fill out for the R Markdown results report. If the experiment is provided, the documentation files can be found in the Experiment's results directory (see Experiment$get_save_dir()) under docs/. Otherwise, the documentation files can be found in the specified save_dir directory under docs/. The documentation files generated include objectives.md and .md files corresponding to DGPs, Methods, Evaluators, and Visualizers in the Experiment.

Usage

create_doc_template(experiment, save_dir)

Arguments

experiment

An Experiment object. If provided, documentation is created for all previously saved Experiments that are found in the directory given by experiment$get_save_dir(). If no Experiments have been previously saved under this directory, then the current experiment is saved to disk and used to create the documentation template.

save_dir

An optional directory in which to find previously saved Experiment objects. Documentation is created for these found Experiments. Not used if experiment is provided.

Value

The original Experiment object if provided. Otherwise, returns NULL.

Examples

if (FALSE) {
# create documentation template from an experiment (of class `Experiment`)
create_doc_template(experiment)

# or alternatively, create documentation template from a specific directory
create_doc_template(save_dir = experiment$get_save_dir())}