Arguments that are shared by multiple Experiment helper funs.

Arguments that are shared by multiple Experiment helper funs.

Arguments

checkpoint_n_reps

The number of experiment replicates to compute before saving results to disk. If 0 (the default), no checkpoints are saved.

dgp

A DGP object.

evaluator

An Evaluator object.

eval_results

A list of result tibbles, as returned by evaluate_experiment().

experiment

An Experiment object.

fit_results

A tibble, as returned by fit_experiment().

future.globals

Character vector of names in the global environment to pass to parallel workers. Passed as the argument of the same name to future.apply::future_lapply and related functions. To set for all runs of the experiment, use the same argument during initialization.

future.packages

Character vector of packages required by parallel workers. Passed as the argument of the same name to future.apply::future_lapply and related functions. To set for all runs of the experiment, use the same argument during initialization.

future.seed

Passed as the argument of the same name in future.apply::future_apply.

method

A Method object.

n_reps

The number of replicates of the Experiment for this run.

parallel_strategy

A vector with some combination of "reps", "dgps", or "methods". Determines how computation will be distributed across available resources. Currently only the default, "reps", is supported.

save

If TRUE, save outputs to disk.

use_cached

Logical. If TRUE, find and return previously saved results. If cached results cannot be found, continue as if use_cached was FALSE.

vary_params

A vector of DGP or Method parameter names that are varied across in the Experiment.

verbose

Level of verbosity. Default is 1, which prints out messages after major checkpoints in the experiment. If 2, prints additional debugging information for warnings and messages from user-defined functions (in addition to error debugging information). If 0, no messages are printed other than user-defined function error debugging information.

visualizer

A Visualizer object.

...

Not used.